Fixed the broken Delphi version of the TileResampler in the Unix branch. Fixed a few other mistakes and cleaned up a bit too.
@Vytas: Really great work on the compression stream stuff, it works perfectly in Lazarus now :) I look forward to you fixing the main game as well so we can have tile textures loading :)
However, when I run the TileResampler on Windows (I haven't tried Linux) I get a blank CMD Window at the same time behind the form. Executing the TilesetResampler.exe or running it from within Lazarus always does this. I was wondering if you knew why, or if it is some sort of debugging that you enabled in this project. The program still works fine, there is just this black window behind it. Does it happen for you on Windows?

A few other notes for you:
 - We try to keep LPR and DPR files identical. If you change one, use compiler directives and copy it to the other. Look at the KaM Remake ones for an example.
 - In order to use the WDC directive you need the .INC file that I added here. It should be included in every unit. If you look at that file you'll see why, it makes any version of Delphi also define WDC. (WDC is not an existing definition)
 - We don't seem to use .RES files in Lazarus, only Delphi. I don't really understand this but all of the other projects only use it in WDC.

I'm not criticising, just explaining because there is no way for you to know these things without being told. Once again thank you very much for fixing the compression stream incompatibilities :)

git-svn-id: svn://localhost/branches/Unix@1695 a386ff7f-d321-c144-bb17-e251d28fee14