Increased the maximum resource in Barracks up to 65k
We did not really had any good reason to limit it to 255 or 32. Cos in first case we just limited it for no reason, player may have as many wares as he wants. And in second case - mapmaker should add wares directly to Barracks, not to Store. And if taking cheats into account - maybe we can make a new cheatcode to add weapons directly to Barracks - would be nice in fact.

git-svn-id: svn://localhost/trunk@2514 a386ff7f-d321-c144-bb17-e251d28fee14