Fixed issues with requiring administrator rights when installed to Program Files. We now have a manifest file which does NOT provide trustInfo, meaning windows will enable file virtualization and let us write into the Virtual Store if executed as non-administrator. Before it was using Delphi's default manifest I think.
Since Windows Vista Program Files is read-only if you are non-administrator. To keep compatibility with old XP applications that write into Program Files, Microsoft invented the "Virtual Store" which allows applications to think they are saving to Program Files through virtualization. If your manifest file declares TrustInfo it will not use the Virtual Store because Microsoft thinks it is an application designed with Program Files as read-only. So by not declaring TrustInfo we can "write" to Program Files even though the data is stored in the virtual store.
HellRacer confirmed this is now fixed when I sent him a test EXE, it successfully wrote files in Program Files (through virtual store) but r3252 did not.

git-svn-id: svn://localhost/trunk@3291 a386ff7f-d321-c144-bb17-e251d28fee14