Updating or adding a dynamic script analogue PlayerColorText
At this stage, I just updated the function of the TextbookText so that it returns not only the clarified color but also the original + so that the function remains compatible with the old scripts.
If you do not like it, I can add a separate one.
If you say that it's not obvious then In the wiki you can update