BugFix: DL maps could be changed and used to start the game, what will cause error on map load
Mostly its script could be changed. And because script can be changed only with external soft, than we have manage this situation as normal for user.

In that case map CRC will be changed and did not match to map filename ending. We have to stop host from selecting this map, but let him see it in map editor list to make it possible 'Move out from downloads' or open and resave it. 
Otherwise user will be forced to make map files (script mainly) absolutely same, as it was at the moment of downloading, which is not easy to archive sometimes. Or manually rename and move all the map files to MapsMP folder, which not convinient.